move app_args & tls to lib
