+++ src/cgw-common/Cargo.toml
@@ -24,3 +24,8 @@ derive_more = { version = "0.99.17" }
 eui48 = { version = "1.1.0", features = ["serde"] }
 tonic = { version = "0.11.0" }
 prometheus = { version = "0.13.4", features = ["process"] }
+url = { version = "2.5.2" }
+base64 = { version = "0.22.0" }
+rustls-pki-types = { version = "1.7.0" }
+x509-parser = { version = "0.16.0" }
+rustls-pemfile = { version = "2.1.2" }
