+++ src/cgw_db_accessor.rs
@@ -1,10 +1,12 @@
-use crate::cgw_app_args::CGWDBArgs;
+use cgw_common::{
+    cgw_app_args::CGWDBArgs,
+    cgw_errors::{Error, Result},
+    cgw_tls::cgw_tls_create_db_connect,
+};
 
-use crate::cgw_tls::cgw_tls_create_db_connect;
 use crate::{
     cgw_metrics::{CGWMetrics, CGWMetricsHealthComponent, CGWMetricsHealthComponentStatus},
 };
-use cgw_common::cgw_errors::{Error, Result};
 
 use eui48::MacAddress;
